# config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "BinairApi",
    "description": "API Authentikasi & User",
    "contact": {}
  },
  "host": "binair-backend-production.up.railway.app",
  "basePath": "/api/v1",
  "securityDefinitions": {},
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/tickets": {
      "get": {
        "summary": "Get All Tickets",
        "tags": ["Tickets"],
        "operationId": "GetAllTickets",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "type": "string",
            "description": ""
          },
          {
            "name": "airport_from",
            "in": "query",
            "type": "string",
            "description": ""
          },
          {
            "name": "to",
            "in": "query",
            "type": "string",
            "description": ""
          },
          {
            "name": "airport_to",
            "in": "query",
            "type": "string",
            "description": ""
          },
          {
            "name": "date_start",
            "in": "query",
            "type": "string",
            "description": ""
          },
          {
            "name": "date_end",
            "in": "query",
            "type": "string",
            "description": ""
          },
          {
            "name": "type",
            "in": "query",
            "type": "string",
            "description": "fill by 'roundtrip' or 'oneway'"
          },
          {
            "name": "willFly",
            "in": "query",
            "type": "string",
            "description": "fill by 'true' or 'false'"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string",
            "description": "which page start by 0, example = 2"
          },
          {
            "name": "size",
            "in": "query",
            "type": "string",
            "description": "size page, example = 10"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "post": {
        "summary": "Create Ticket",
        "tags": ["Tickets"],
        "operationId": "CreateTicket",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateTicketRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/tickets/{id-p}": {
      "put": {
        "summary": "Update Ticket",
        "tags": ["Tickets"],
        "operationId": "UpdateTicket",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "id-p",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateTicketRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/tickets/{id-d}": {
      "delete": {
        "summary": "Delete Ticket",
        "tags": ["Tickets"],
        "operationId": "DeleteTicket",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "id-d",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/wishlists": {
      "get": {
        "summary": "Get All Wishlist",
        "tags": ["Wishlists"],
        "operationId": "GetAllWishlist",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string",
            "description": "which page start by 0, example = 2"
          },
          {
            "name": "size",
            "in": "query",
            "type": "string",
            "description": "size page, example = 10"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "Create Wishlist",
        "tags": ["Wishlists"],
        "operationId": "CreateWishlist",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateWishlistRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/wishlists/{id-d}": {
      "delete": {
        "summary": "Delete Wishlist",
        "tags": ["Wishlists"],
        "operationId": "DeleteWishlist",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "id-d",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateTicketRequest": {
      "title": "CreateTicketRequest",
      "example": {
        "from": "Jakarta",
        "to": "Bali",
        "airport_from": "Soekarno Hatta International Airport",
        "airport_to": "I Gusti Ngurah Rai International Airport",
        "departure_time": "14:00",
        "arrival_time": "16:30",
        "date_start": "22-01-2023",
        "date_end": "25-01-2023",
        "type": "roundtrip",
        "adult_price": 2200000,
        "child_price": 1500000,
        "available": true,
        "init_stock": 200,
        "curr_stock": 200
      },
      "type": "object",
      "properties": {
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        },
        "airport_from": {
          "type": "string"
        },
        "airport_to": {
          "type": "string"
        },
        "departure_time": {
          "type": "string"
        },
        "arrival_time": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "adult_price": {
          "type": "integer",
          "format": "int32"
        },
        "child_price": {
          "type": "integer",
          "format": "int32"
        },
        "available": {
          "type": "boolean"
        },
        "init_stock": {
          "type": "integer",
          "format": "int32"
        },
        "curr_stock": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "from",
        "to",
        "airport_from",
        "airport_to",
        "departure_time",
        "arrival_time",
        "date",
        "type",
        "adult_price",
        "child_price",
        "available",
        "init_stock",
        "curr_stock"
      ]
    },
    "UpdateTicketRequest": {
      "title": "UpdateTicketRequest",
      "example": {
        "type": "oneway",
        "date_end": null
      },
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        }
      },
      "required": ["type"]
    },
    "CreateWishlistRequest": {
      "title": "CreateWishlistRequest",
      "example": {
        "usersId": "9dde98a9-8a7e-4139-95a6-1182fbccdac0",
        "ticketsId": "ced2a7f6-6096-4e4b-99f3-d890c238f5f2"
      },
      "type": "object",
      "properties": {
        "usersId": {
          "type": "string"
        },
        "ticketsId": {
          "type": "string"
        }
      },
      "required": ["usersId", "ticketsId"]
    },
...
  },
  "security": [],
...
}